refactor(providers): use dynamic import for subcommand
