+++ packages/opencode/src/index.ts
@@ -4,7 +4,7 @@ import { RunCommand } from "./cli/cmd/run"
 import { GenerateCommand } from "./cli/cmd/generate/command"
 import * as Log from "@opencode-ai/core/util/log"
 import { ConsoleCommand } from "./cli/cmd/account/command"
-import { ProvidersCommand } from "./cli/cmd/providers"
+import { ProvidersCommand } from "./cli/cmd/providers/command"
 import { AgentCommand } from "./cli/cmd/agent/command"
 import { UpgradeCommand } from "./cli/cmd/upgrade"
 import { UninstallCommand } from "./cli/cmd/uninstall"
